Final upload of fully working project
Fully working project with file open, file read and file close, as well as creating a tree, encoding and decoding based on Huffman's algorithm for compression/decompression. To change the input of the program, you must write a different string inside of "file.txt". The probability of a certain char occurrence is being calculated as well.